Defect for delete of preinstalled directory (#20)
* defect for delete of preinstalled directory

* updating deployweb

* updating repo split for deployjs

* code cleanup updates

+++ packages/actions/lib/common.js
@@ -1,16 +1,11 @@
 const fs = require('fs');
 const path = require('path');
-const exec = require('child_process').exec;
-const git = require('simple-git');
-const yaml = require('js-yaml');
+const { exec } = require('child_process');
+
 let command = '';
 
 /**
- * Action to deploy openwhisk elements from a compliant repository
- *  @param {string} gitUrl - github url containing the manifest and elements to deploy
- *  @param {string} manifestPath - (optional) the path to the manifest file, e.g. "openwhisk/src"
- *  @param {object} envData - (optional) some specific details such as cloudant username or cloudant password
- *  @return {object} Promise
+ * Common function to enable deployment from deployWeb.js & deploy.js
  */
 function main(params) {
   return new Promise((resolve, reject) => {
@@ -18,6 +13,7 @@ function main(params) {
       wskAuth,
       wskApiHost,
       manifestPath,
+      usingTemp,
       manifestFileName,
       repoDir,
       envData,
@@ -40,13 +36,17 @@ function main(params) {
 
     const manifestFilePath = `${repoDir}/${manifestPath}/${manifestFileName}`;
     if (!fs.existsSync(manifestFilePath)) {
-      deleteFolder(repoDir);
-      reject(`Error loading manifest file. Does a manifest file exist?`);
+      if (usingTemp) {
+        deleteFolder(repoDir);
+      }
+      reject(new Error('Error loading manifest file. Does a manifest file exist?'));
     } else {
       exec(command, execOptions, (err, stdout, stderr) => {
-        deleteFolder(repoDir);
+        if (usingTemp) {
+          deleteFolder(repoDir);
+        }
         if (err) {
-          reject('Error running `./wskdeploy`: ', err);
+          reject(new Error('Error running `./wskdeploy`: ', err));
         }
         if (stdout) {
           console.log('stdout from wskDeploy: ', stdout, ' type ', typeof stdout);
@@ -62,7 +62,7 @@ function main(params) {
           if (typeof stdout === 'object') {
             if (stdout.error) {
               stdout.descriptiveError = 'Could not successfully run wskdeploy. Please run again with the verbose flag, -v.';
-              reject(stdout);
+              reject(new Error(stdout));
             }
           }
         }
@@ -86,12 +86,12 @@ function main(params) {
  */
 function deleteFolder(pathToDelete) {
   if (fs.existsSync(pathToDelete)) {
-    fs.readdirSync(pathToDelete).forEach(function(file, index){
-      var curPath = path.join(pathToDelete, file);
+    fs.readdirSync(pathToDelete).forEach((file, index) => {
+      const curPath = path.join(pathToDelete, file);
       if (fs.lstatSync(curPath).isDirectory()) {
         deleteFolder(curPath);
       } else {
-        //unlinkSync deletes files.
+        // unlinkSync deletes files.
         fs.unlinkSync(curPath);
       }
     });
@@ -100,5 +100,5 @@ function deleteFolder(pathToDelete) {
 }
 
 module.exports = {
-  'main': main
+  main,
 };
